Cherry-picks to stable for 1.18+ck2 (#1)
* Remove allow_hosts workaround for Focal and after (#169)

* Fix regression for offline deployments (#172)

Add back the writing of ``/root/.pydistutils.cfg`` for all series
and drop the now removed ``allow_hosts`` keyword for Eoan and
onwards.

The ``find_links`` keyword with path to wheelhouse still needs to
be there for successful offline install of some Python packages.

Fixes #171

* Pin netaddr (#179)

netaddr 0.7.20 pulls in importlib-resources, which depends on a newer
version of setuptools_scm than we use.  It also pulls zipp, which
depends on a newer version of setuptools than we use.  Pinning to
0.7.19 avoids this.

* Stop pip3 AND easy_install from accessing network on install (#184)

* Stop pip3 AND easy_install from accessing network on install

This change prevents the wheelhouse installation from accessing the
network at all.  This allows it to install quickly behind firewalls that
silently drop packets causing long delays in installing the packages if
a dependency is 'later' in the wheelhouse install, which is done in
directory order (usually alphabetically).

Closes launchpad bug: #1884449

* Add required snap install core line

* Remove Eoan from functional tests (#186)

Eoan has been EOLed and no longer has images available, so it gets stuck
in pending and causes the functional tests to hang.

* Fix lint failures (from #170)

* Use ``travis_wait`` to keep func test alive (#162)

At present we have enabled excessive debug-logging to accompilsh
the same thing, but that makes it unpleasant to look at.

Change permissions on the LXD socket to give travis user access
instead of using the sudo back to ourself trick as it leads to
some unwanted complexity.

Add Python 3.8 target to test matrix.

Replace ``cryptography`` with ``psycopg2`` as it has not banned
Python 3.4 support.

Closes #161

Co-authored-by: Frode Nordahl <frode.nordahl@canonical.com>
Co-authored-by: Paul Goins <paul.goins@canonical.com>
Co-authored-by: Alex Kavanagh <567675+ajkavanagh@users.noreply.github.com>